+++ packages/render/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @computesdk/render
 
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [3333388]
+  - computesdk@1.13.0
+  - @computesdk/provider@1.0.9
+
+## 1.1.14
+
+### Patch Changes
+
+- Updated dependencies [4decff7]
+  - @computesdk/provider@1.0.8
+  - computesdk@1.12.1
+
+## 1.1.13
+
+### Patch Changes
+
+- Updated dependencies [fdda069]
+  - computesdk@1.12.0
+  - @computesdk/provider@1.0.7
+
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies [7c8d968]
+  - computesdk@1.11.1
+  - @computesdk/provider@1.0.6
+
 ## 1.1.11
 
 ### Patch Changes